refactor: 훅, 컴포넌트로 TravelKanban 분리
