+++ app/(tabs)/schedule/index.tsx
@@ -1,7 +1,7 @@
 import ButtonSection from '@/conponents/(tabs)/schedule/ButtonSection';
 import ScheduleTopBar from '@/conponents/(tabs)/schedule/ScheduleTopBar';
 import Screen from '@/conponents/(tabs)/schedule/screens';
-import { FloatingPortalProvider } from '@/conponents/(tabs)/schedule/screens/PlanScreen/FloatingPortal';
+import { FloatingPortalProvider } from '@/conponents/(tabs)/schedule/screens/EditPlanScreen/TravelKanban/FloatingPortal';
 import Stepper from '@/conponents/(tabs)/schedule/Stepper';
 import React from 'react';
 import { StyleSheet, View } from 'react-native';
