+++ conponents/(tabs)/schedule/screens/EditPlanScreen/dummyData.ts
@@ -1,0 +1,88 @@
+export const dummySchedule = [
+    [
+      {
+        "id": 9007199254740991,
+        "tourismType": "ATTRACTION_SPOT",
+        "name": "경복궁",
+        "fullAddress": "서울특별시 종로구 사직로 161",
+        "longitude": 126.9770,
+        "latitude": 37.5796,
+        "phoneNumber": "02-3700-3900",
+        "imgUrl": "https://picsum.photos/seed/gyeongbokgung/400/300"
+      },
+      {
+        "id": 9007199254740992,
+        "tourismType": "RESTAURANT",
+        "name": "토속촌 삼계탕",
+        "fullAddress": "서울특별시 종로구 자하문로5길 5",
+        "longitude": 126.9721,
+        "latitude": 37.5795,
+        "phoneNumber": "02-737-7444",
+        "imgUrl": "https://picsum.photos/seed/tosokchon/400/300"
+      },
+      {
+        "id": 9007199254740993,
+        "tourismType": "ATTRACTION_SPOT",
+        "name": "N서울타워",
+        "fullAddress": "서울특별시 용산구 남산공원길 105",
+        "longitude": 126.9882,
+        "latitude": 37.5512,
+        "phoneNumber": "02-3455-9277",
+        "imgUrl": "https://picsum.photos/seed/nseoultower/400/300"
+      }
+    ],
+    [
+      {
+        "id": 9007199254740994,
+        "tourismType": "ATTRACTION_SPOT",
+        "name": "해운대 해수욕장",
+        "fullAddress": "부산광역시 해운대구 우동",
+        "longitude": 129.1586,
+        "latitude": 35.1581,
+        "phoneNumber": "051-749-4335",
+        "imgUrl": "https://picsum.photos/seed/haeundae/400/300"
+      },
+      {
+        "id": 9007199254740995,
+        "tourismType": "RESTAURANT",
+        "name": "해운대암소갈비집",
+        "fullAddress": "부산광역시 해운대구 중동2로10번길 32-10",
+        "longitude": 129.1634,
+        "latitude": 35.1611,
+        "phoneNumber": "051-746-0033",
+        "imgUrl": "https://picsum.photos/seed/haewundae-galbi/400/300"
+      }
+    ],
+    [
+      {
+        "id": 9007199254740996,
+        "tourismType": "ATTRACTION_SPOT",
+        "name": "성산일출봉",
+        "fullAddress": "제주특별자치도 서귀포시 성산읍 성산리 1",
+        "longitude": 126.9423,
+        "latitude": 33.4581,
+        "phoneNumber": "064-783-0959",
+        "imgUrl": "https://picsum.photos/seed/sungsan/400/300"
+      },
+      {
+        "id": 9007199254740997,
+        "tourismType": "RESTAURANT",
+        "name": "제주 흑돼지 맛집 '돈사돈'",
+        "fullAddress": "제주특별자치도 제주시 우평로 19",
+        "longitude": 126.4786,
+        "latitude": 33.4851,
+        "phoneNumber": "064-746-8989",
+        "imgUrl": "https://picsum.photos/seed/donsadon/400/300"
+      },
+      {
+        "id": 9007199254740998,
+        "tourismType": "CAFE",
+        "name": "카페 봄날",
+        "fullAddress": "제주특별자치도 제주시 애월읍 애월로1길 25",
+        "longitude": 126.3106,
+        "latitude": 33.4939,
+        "phoneNumber": "064-799-4999",
+        "imgUrl": "https://picsum.photos/seed/bomnalcafe/400/300"
+      }
+    ]
+  ]
