+++ conponents/(tabs)/schedule/screens/index.tsx
@@ -4,9 +4,9 @@ import { ScreenKey } from "@/types";
 import { useShallow } from "zustand/shallow";
 import CalanderScreen from "./CalanderScreen";
 import CompleteScreen from "./CompleteScreen";
+import EditPlanScreen from "./EditPlanScreen";
 import GenerateScheduleScreen from "./GenerateScheduleScreen";
 import AIRecommendScreen from "./PlanScreen/AIRecommendPlanScreen";
-import EditPlanScreen from "./PlanScreen/EditPlanScreen";
 import SelectActivityScreen from "./SelectPlaceScreen/SelectActivityScreen";
 import SelectPlaceScreen from "./SelectPlaceScreen/SelectPlaceScreen";
 import SelectRegionScreen from "./SelectRegionScreen";
